+++ ricaun.NUnit/Services/TestAssemblyService.cs
@@ -1,5 +1,4 @@
-﻿using NUnit.Framework;
-using ricaun.NUnit.Extensions;
+﻿using ricaun.NUnit.Extensions;
 using ricaun.NUnit.Models;
 using System;
 using System.Collections.Generic;
@@ -51,17 +50,17 @@ namespace ricaun.NUnit.Services
         /// <summary>
         /// Get Test Types
         /// </summary>
+        /// <param name="useExported">A flag indicating whether to use exported types only.</param>
         /// <returns></returns>
-        private IEnumerable<Type> GetTestTypes()
+        private IEnumerable<Type> GetTestTypes(bool useExported = false)
         {
             var types = new List<Type>();
 
-            foreach (Type type in this.assembly.GetTypes())
+            foreach (Type type in GetAssemblyTypes(useExported))
             {
                 if (!type.IsClass) continue;
                 if (type.IsAbstract) continue;
 
-                //if (GetFilterTestMethods(type).Any())
                 if (AnyMethodWithTestAttribute(type))
                 {
                     types.Add(type);
@@ -70,24 +69,25 @@ namespace ricaun.NUnit.Services
             return types.OrderBy(e => e.FullName);
         }
 
-        ///// <summary>
-        ///// GetTestTypeMethods
-        ///// </summary>
-        ///// <returns></returns>
-        //[Obsolete("Does not work with Abstract")]
-        //public IEnumerable<MethodInfo> GetTestTypeMethods()
-        //{
-        //    return GetTestTypes().SelectMany(e => e.GetMethods().Where(AnyTestAttribute)).OrderBy(e => GetMethodFullName(e));
-        //}
+        /// <summary>
+        /// Get the types from the specified assembly.
+        /// </summary>
+        /// <param name="useExported">A flag indicating whether to use exported types only.</param>
+        /// <returns>An array of types from the assembly.</returns>
+        private Type[] GetAssemblyTypes(bool useExported)
+        {
+            return useExported ? this.assembly.GetExportedTypes() : this.assembly.GetTypes();
+        }
 
         /// <summary>
         /// GetTestDictionaryTypeMethods
         /// </summary>
+        /// <param name="useExported">A flag indicating whether to use exported types only.</param>
         /// <returns></returns>
-        public IReadOnlyDictionary<Type, MethodInfo[]> GetTestDictionaryTypeMethods()
+        public IReadOnlyDictionary<Type, MethodInfo[]> GetTestDictionaryTypeMethods(bool useExported = false)
         {
             var result = new Dictionary<Type, MethodInfo[]>();
-            var types = GetTestTypes();
+            var types = GetTestTypes(useExported);
             foreach (var type in types)
             {
                 var methods = GetMethodWithTestAttribute(type).ToArray();
@@ -102,10 +102,11 @@ namespace ricaun.NUnit.Services
         /// <summary>
         /// GetTestFullNames
         /// </summary>
+        /// <param name="useExported">A flag indicating whether to use exported types only.</param>
         /// <returns></returns>
-        public IEnumerable<string> GetTestFullNames()
+        public IEnumerable<string> GetTestFullNames(bool useExported = false)
         {
-            foreach (var typeMethod in GetTestDictionaryTypeMethods())
+            foreach (var typeMethod in GetTestDictionaryTypeMethods(useExported))
             {
                 Type type = typeMethod.Key;
                 MethodInfo[] methods = typeMethod.Value;
@@ -119,24 +120,15 @@ namespace ricaun.NUnit.Services
             }
         }
 
-        ///// <summary>
-        ///// GetTestTypeMethods
-        ///// </summary>
-        ///// <param name="fullName"></param>
-        ///// <returns></returns>
-        //public MethodInfo GetTestTypeMethods(string fullName)
-        //{
-        //    return GetTestTypeMethods().FirstOrDefault(e => GetMethodFullName(e) == fullName);
-        //}
-
         /// <summary>
-        /// Test
+        /// RunTests
         /// </summary>
+        /// <param name="useExported">A flag indicating whether to use exported types only.</param>
         /// <returns></returns>
-        public IEnumerable<TestTypeModel> Test()
+        public IEnumerable<TestTypeModel> RunTests(bool useExported = false)
         {
             var result = new List<TestTypeModel>();
-            var types = GetTestTypes().Where(AnyMethodWithTestAttributeAndFilter);
+            var types = GetTestTypes(useExported).Where(AnyMethodWithTestAttributeAndFilter);
 
             foreach (var type in types)
             {