+++ ricaun.NUnit/TestEngine.cs
@@ -3,6 +3,7 @@ using ricaun.NUnit.Extensions;
 using ricaun.NUnit.Models;
 using ricaun.NUnit.Services;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -25,6 +26,7 @@ namespace ricaun.NUnit
         /// <summary>
         /// Initialize
         /// </summary>
+        /// <param name="message">The initialization message</param>
         /// <returns>Return true if successful</returns>
         public static bool Initialize(out string message)
         {
@@ -46,9 +48,10 @@ namespace ricaun.NUnit
         /// </summary>
         /// <param name="assemblyFile">Assembly location</param>
         /// <param name="parameters">Parameters to create classes and methods</param>
-        /// <returns></returns>
+        /// <returns>The TestAssemblyModel object</returns>
         public static TestAssemblyModel TestAssembly(string assemblyFile, params object[] parameters)
         {
+            Exceptions.Clear();
             var testAssemblyModel = new TestAssemblyModel();
             using (var console = new ConsoleWriterDateTime())
             {
@@ -70,14 +73,32 @@ namespace ricaun.NUnit
 
                         ValidateTestAssemblyNUnitVersion(testAssembly.Assembly);
 
-                        var tests = testAssembly.Test();
+                        try
+                        {
+                            testAssemblyModel.Tests.AddRange(testAssembly.RunTests());
+                        }
+                        catch (Exception ex)
+                        {
+                            Exceptions.Add(ex);
+                            Debug.WriteLine(ex);
 
-                        testAssemblyModel.Tests.AddRange(tests);
+                            try
+                            {
+                                testAssemblyModel.Tests.AddRange(testAssembly.RunTests(true));
+                            }
+                            catch (Exception ex2)
+                            {
+                                Exceptions.Add(ex2);
+                                Debug.WriteLine(ex2);
+                            }
+                        }
+
                         testAssemblyModel.Success = !testAssemblyModel.Tests.Any(e => !e.Success);
                     }
                 }
                 catch (Exception ex)
                 {
+                    Exceptions.Add(ex);
                     if (string.IsNullOrEmpty(testAssemblyModel.Name))
                         testAssemblyModel.Name = ex.GetType().Name;
 
@@ -91,20 +112,41 @@ namespace ricaun.NUnit
         }
 
         /// <summary>
+        /// Exceptions for the GetTestFullNames
+        /// </summary>
+        public static IList<Exception> Exceptions { get; } = new List<Exception>();
+
+        /// <summary>
         /// GetTestFullNames
         /// </summary>
-        /// <param name="assemblyFile"></param>
-        /// <returns></returns>
+        /// <param name="assemblyFile">The assembly location</param>
+        /// <returns>An array of test full names</returns>
         public static string[] GetTestFullNames(string assemblyFile)
         {
+            Exceptions.Clear();
             using (new AssemblyResolveCurrentDirectoryService(Path.GetDirectoryName(assemblyFile)))
             {
                 try
                 {
                     var testAssembly = new TestAssemblyService(assemblyFile);
-                    return testAssembly.GetTestFullNames().ToArray();
+                    try
+                    {
+                        return testAssembly.GetTestFullNames().ToArray();
+                    }
+                    catch (Exception ex)
+                    {
+                        Exceptions.Add(ex);
+                        Debug.WriteLine(ex);
+                    }
+
+                    // Retry with exported types only
+                    return testAssembly.GetTestFullNames(true).ToArray();
                 }
-                catch (Exception ex) { Debug.WriteLine(ex); }
+                catch (Exception ex)
+                {
+                    Exceptions.Add(ex);
+                    Debug.WriteLine(ex);
+                }
             }
             return new string[] { };
         }
@@ -112,10 +154,10 @@ namespace ricaun.NUnit
         /// <summary>
         /// GetTestFullNames
         /// </summary>
-        /// <param name="assemblyFile"></param>
-        /// <param name="directoryResolver"></param>
-        /// <param name="includeSubDirectories"></param>
-        /// <returns></returns>
+        /// <param name="assemblyFile">The assembly location</param>
+        /// <param name="directoryResolver">The directory resolver</param>
+        /// <param name="includeSubDirectories">Flag to include subdirectories</param>
+        /// <returns>An array of test full names</returns>
         public static string[] GetTestFullNames(string assemblyFile, string directoryResolver, bool includeSubDirectories = false)
         {
             using (new AssemblyResolveService(directoryResolver, includeSubDirectories))
@@ -125,12 +167,11 @@ namespace ricaun.NUnit
         }
 
         /// <summary>
-        /// GetTestFullName
-        /// <code>TypeName.TestName.TestAlias</code>
+        /// Get the full name of a test
         /// </summary>
-        /// <param name="testType"></param>
-        /// <param name="test"></param>
-        /// <returns></returns>
+        /// <param name="testType">The test type model</param>
+        /// <param name="test">The test model</param>
+        /// <returns>The full name of the test</returns>
         [Obsolete("This method is obsolete, use TestModel.FullName")]
         internal static string GetTestFullName(TestTypeModel testType, TestModel test)
         {
@@ -149,26 +190,21 @@ namespace ricaun.NUnit
         public static Version VersionNUnit => typeof(NUnitAttribute).Assembly.GetName().Version;
 
         /// <summary>
-        /// Check <paramref name="assemblyFile"/> assembly contain NUnit reference
+        /// Check if the assembly contains a reference to NUnit
         /// </summary>
-        /// <remarks>
-        /// This method use <seealso cref="Assembly.ReflectionOnlyLoadFrom"/> 
-        /// </remarks>
-        /// <param name="assemblyFile"></param>
-        /// <returns></returns>
+        /// <param name="assemblyFile">The assembly location</param>
+        /// <returns>True if the assembly contains a reference to NUnit, otherwise false</returns>
         public static bool ContainNUnit(string assemblyFile)
         {
             return GetReferencedAssemblyNUnit(assemblyFile) is not null;
         }
+
         /// <summary>
-        /// Check <paramref name="assemblyFile"/> assembly contain NUnit reference
+        /// Check if the assembly contains a reference to NUnit
         /// </summary>
-        /// <remarks>
-        /// This method use <seealso cref="Assembly.ReflectionOnlyLoadFrom"/> 
-        /// </remarks>
-        /// <param name="assemblyFile"></param>
-        /// <param name="assemblyName"></param>
-        /// <returns></returns>
+        /// <param name="assemblyFile">The assembly location</param>
+        /// <param name="assemblyName">The assembly name</param>
+        /// <returns>True if the assembly contains a reference to NUnit, otherwise false</returns>
         public static bool ContainNUnit(string assemblyFile, out AssemblyName assemblyName)
         {
             assemblyName = GetReferencedAssemblyNUnit(assemblyFile);
@@ -179,17 +215,12 @@ namespace ricaun.NUnit
         #region nunit.framework
         private const string NUNIT_ASSEMBLY = "nunit.framework";
         /// <summary>
-        /// Get Referenced NUnit using <seealso cref="Assembly.ReflectionOnlyLoadFrom"/>
+        /// Get the referenced NUnit assembly using Assembly.ReflectionOnlyLoadFrom
         /// </summary>
-        /// <param name="assemblyFile"></param>
-        /// <returns></returns>
+        /// <param name="assemblyFile">The assembly location</param>
+        /// <returns>The referenced NUnit assembly</returns>
         private static AssemblyName GetReferencedAssemblyNUnit(string assemblyFile)
         {
-            //var assembly = ReflectionOnlyLoadFrom(assemblyFile);
-
-            //if (assembly is null) return null;
-            //AssemblyName nunitReference = GetNUnitReference(assembly);
-
             var references = ReferenceLoaderUtils.GetReferencedAssemblies(assemblyFile);
 
             var nunitReference = references.FirstOrDefault(e => e.Name.Equals(NUNIT_ASSEMBLY));
@@ -203,6 +234,7 @@ namespace ricaun.NUnit
                 .FirstOrDefault(e => e.Name.Equals(NUNIT_ASSEMBLY));
         }
 
+        [Obsolete]
         private static Assembly ReflectionOnlyLoadFrom(string assemblyFile)
         {
             try
@@ -227,7 +259,7 @@ namespace ricaun.NUnit
 
             if (nunitReference.Version != VersionNUnit)
             {
-                var message = $"'{NUNIT_ASSEMBLY}' version {nunitReference.Version.ToString(3)} is not allow. Use the version {VersionNUnit.ToString(3)}.";
+                var message = $"'{NUNIT_ASSEMBLY}' version {nunitReference.Version.ToString(3)} is not allowed. Use the version {VersionNUnit.ToString(3)}.";
                 Debug.WriteLine(message);
                 throw new FileLoadException(message);
             }